jump label: Introduce very_likely()
The current very_unlikely() construct, assumes that the branch
will be disabled by default. This means we have a single no-op
in the straight line path, and when the branch is made true we
patch the no-op with a jump.

There are cases (sched feat code), where we want the branch to
default to true, so that the straight line code is the true
branch, and the false branch is enabled via a jump.

In order to implement this while having, jump_label_inc(), and
jump_label_dec() retain their current meaning, we have to store
the initial branch state. I'm using the lowest bit of the
'entries' pointer in the jump_label_key struct, since this
points to memory that has to be aligned to the arch pointer
size. We could have stored this in the 'struct jump_entry' data
structure, but I wanted to avoid adding additional space
overhead.

Thus, the new API is initialized as:

	struct jump_label_key true_key = JUMP_LABEL_INIT_TRUE;

			or

	struct jump_label_key false_key = JUMP_LABEL_INIT_FALSE;

Leaving out an initialization, defaults to false, as in:

	struct jump_label_key uninitialized_key;

Then, for the branches we have:

	very_unlikely(&false_key);

		or

	very_likely(&true_key);

And finally, jump_label_inc(&key), jump_label_dec(&key) are
unchanged - 'jump_label_inc()' means 'make true', and
jump_label_dec()' means make false, with the expected
increment/decrement counting.

Thus, you must use 'true_key' with a 'very_likely()' branch, and
a 'false_key' or an 'uninitialized_key' with a 'very_unlikely()'
branch. Mixing different branches with the same jump_label_key
is not allowed.

I've left the old static_branch(), (which is the same as the new
very_unlikely()), at least until we've converted over all
in-tree and pending users.

Signed-off-by: Jason Baron <jbaron@redhat.com>
Cc: rostedt@goodmis.org
Cc: mathieu.desnoyers@efficios.com
Cc: davem@davemloft.net
Cc: ddaney.cavm@gmail.com
[ Simplified code using jump_label_type(); removed CONFIG_MODULE
  special casing, since C99 initialization inits all unmentioned
  members to 0; obfuscated key->entries initialization. ]
Signed-off-by: Peter Zijlstra <a.p.zijlstra@chello.nl>
Link: http://lkml.kernel.org/r/742699c7e855a61f881c501ea415a43a5da9cd04.1324493360.git.jbaron@redhat.com
Signed-off-by: Ingo Molnar <mingo@elte.hu>