perf-probe: Show symbol+offset for address only kprobes
Show the symbol+offset information for address only kprobe
events when --list operation without debuginfo. Currently
those events are shown by the address itself. With this change
perf probe finds symbols on those addresses and shows it.

E.g. without this change (when debuginfo is not available);
# ./perf probe -l
  probe:t_show         (on 0xffffffff810f6dd0 with m v)
  probe:t_show_1       (on 0xffffffff81100340 with m v t)
  probe:t_show_2       (on 0xffffffff8110a4f0 with m v fmt)
  probe:t_show_3       (on 0xffffffff8110c160 with m v)
  probe:t_show_4       (on 0xffffffff81112470 with m v file)

With this change;
# ./perf probe -l
  probe:t_show         (on t_show with m v)
  probe:t_show_1       (on t_show with m v t)
  probe:t_show_2       (on t_show with m v fmt)
  probe:t_show_3       (on t_show with m v)
  probe:t_show_4       (on t_show with m v file)