module: Lock-down a module when loading with a flag
Lock-down a module so that it is not removed unless forcibly
unload. This is done by loading a module with MODULE_INIT_LOCKDOWN
flag (via finit_module).
This speeds up try_module_get by skipping refcount inc/dec for
the lockdown modules.

Note that this flag requires to update libkmod to support it.

Signed-off-by: Masami Hiramatsu <masami.hiramatsu@gmail.com>