perf record: Support recording SDT events
perf record -e "sdt_PROVIDER:EVENT" is available as same as
other tracepoint events.
In this version, the wildcard and @filename or build-id are not
supported yet.

Note that this doesn't clear (unregister) probe event after recorded
events, since that can skip to register same event next time and
anyway we can not unregister it if someone use the event via ftrace.

Signed-off-by: Masami Hiramatsu <masami.hiramatsu.pt@hitachi.com>
Signed-off-by: Masami Hiramatsu <mhiramat@kernel.org>
---
 Changes in v4:
  - remove internal_call parameter since we don't need suppress messages.