hostfs: Add getattr operation to update inode attribute
Add getattr inode operation to update inode attribute
which can be modified in host-side.

This hostfs_getattr() considers 2 cases, one is for
the files which is not opened yet, and other is for
the files which is already opened. For former case,
it just does lstat(2) with the name, but the latter
case, it has to use opened fd with fstat(2), because
the target file can be renamed in host side.

Signed-off-by: Masami Hiramatsu <mhiramat@kernel.org>