hostfs: Add read_iter ops to follow up host-side update
Add read_iter file operation for following up the update
made in host-side. Since the readpage ops can be invoked
only on the pages which are not read(cached) yet, it can
not work for the pages which has been updated by host-side.

It is able to read such update if the hostfs has its own
read_iter operation. Of couse this may degradate the
performance because it always read the file from host-side.

Signed-off-by: Masami Hiramatsu <mhiramat@kernel.org>