genirq: Implement bitmap matrix allocator
Implement the infrastructure for a simple bitmap based allocator, which
will replace the x86 vector allocator.

It's in the core code as other architectures might be able to reuse/extend
it.

For now it only implements allocations for single CPUs, but it's simple to
add multi CPU allocation support if required.

Signed-off-by: Thomas Gleixner <tglx@linutronix.de>
Cc: Chris Metcalf <cmetcalf@mellanox.com>