x86/vector: Add domain specific debugfs output
Add the debug callback for the vector domain, which gives a detailed
information about vector usage if invoked for the domain by using rhe
matrix allocator debug function and vector/target information when invoked
for a particular interrupt.

Extra information foir the Vector domain:

 CPU  : resvd: alloc: vectors
     0:    39:    25: 0-19,32-42,48-63,128,239-254
     1:    39:     0: 0-19,32,50,128,239-254

Extra information for interrupts:

     Vector:    42
     Target:     4

This allows a detailed analysis of the vector usage and the association to
interrupts and devices.

Signed-off-by: Thomas Gleixner <tglx@linutronix.de>