x86/vector: Simplify cleanup vector handling
The cleanup vector handling iterates over the full vector space, which is
bogus to begin with because cleanup vectors can only be located in the
external vector space. But it's nevertheless a lot of work and cache lines
in hard interrupt context just to find ususally a single vector which needs
to be cleaned up.

With the single CPU affinity restriction all this can be replaced by a per
cpu list. Before the cleanup IPI is sent, the apic data is queued into the
per cpu list of the previous target CPU.

The cleanup handler iterates its per cpu list, does the necessary sanity
checks and either sends a self IPI when the IRR bit of the vector is
pending, or removes the data from the list and cleans up the vector.

This is completely serialized by vector_lock and all race condition
prevention checks can be removed. The data is either on the list and valid
or it has been removed before the cleanup handler runs.

Signed-off-by: Thomas Gleixner <tglx@linutronix.de>