x86/irq: Simplify hotplug vector accounting
Before a CPU is taken offline the number of active interrupt vectors on the
outgoing CPU and the number of vectors which are available on the other
online CPUs are counted and compared. If the active vectors are more than
the available vectors on the other CPUs then the CPU hot-unplug operation
is aborted.

This again uses loop based search. Simplify it by using the information in
the bitmap matrix allocator and only do a bitmap assisted detailed check on
irq descriptors when the initial estimation claims that not enough vectors
are available.

Signed-off-by: Thomas Gleixner <tglx@linutronix.de>