kprobes/x86: Support nested kprobes
For the normal kprobes, there is no chance to nest another
kprobes, because the interrupt is disabled while processing
kprobes (including single stepping.) However, for kprobes
based on ftrace or optimized by jump are now interruptible.
This means another probe on interrupt path can be hit
while those kprobes are processed.

Current kprobes prohibits nested kprobe because that means
such kprobes are called from kprobe handler and which can
cause the infinit loop of int3 and stop the system.

This changes it by introducing push/pop current kprobe and
control block so that nesting kprobe is allowed in a few
level.

Signed-off-by: Masami Hiramatsu <mhiramat@kernel.org>