tracing/kprobe: bpf: Check error injectable event is on function entry
Check whether error injectable event is on function entry or not.
Currently it checks the event is ftrace-based kprobes or not,
but that is wrong. It should check if the event is on the entry
of target function. Since error injection will override a function
to just return with modified return value, that operation must
be done before the target function starts making stackframe.

As a side effect, bpf error injection is no need to depend on
function-tracer. It can work with sw-breakpoint based kprobe
events too.

Signed-off-by: Masami Hiramatsu <mhiramat@kernel.org>
---
  Changes in v3:
   - Move arch_ftrace_kprobe_override_function() to
     core.c because it is no longer depending on ftrace.
   - Fix a bug to skip passing kprobes target name to
     kprobe_on_func_entry(). Passing both @addr and @symbol
     to that function will result in failure.