tracing: Simplify creation and deletion of synthetic event
Simplify creation and deletion code of synthetic event.

Since the event_mutex and synth_event_mutex ordering issue
is gone, we can skip existing event check when adding or
deleting event, and some redundant code in error path.

This changes release_all_synth_events() to abort the process
when it hits any error and returns the error code. It succeeds
only if it has no error.

Signed-off-by: Masami Hiramatsu <mhiramat@kernel.org>