perf/probe: Generate event name with line number
Generate event name from function name with line number
as <function>_L<line_number>. Note that this is only for
the new event which is defined by function and lines.

If there is another event on same line, you have to use
"-f" option. In that case, the new event has "_1" suffix.

 e.g.
 # perf probe -l
  probe:kernel_read_L1 (on kernel_read:1@ksrc/linux/fs/read_write.c)
 # perf probe -fa kernel_read:1
  Added new event:
    probe:kernel_read_L1_1 (on kernel_read:2)

Signed-off-by: Masami Hiramatsu <mhiramat@kernel.org>