perf/probe: Walk function lines in lexical blocks
Since some inlined functions are in lexical blocks of given
function, we have to recursively walk through the DIE tree.
Without this fix, perf-probe -L can miss the inlined functions
which is in a lexcical block (like if (..) { func() } case.)

However, even though, to walk the lines in a given function,
we don't need to follow the children DIE of inlined functions
because those do not have any lines in the specified function.

We need to walk though whole trees only if we walk all lines
in a given file, because an inlined function can include
another inlined function in the same file.

Fixes: b0e9cb2802d4 ("perf probe: Fix to search nested inlined functions in CU")
Signed-off-by: Masami Hiramatsu <mhiramat@kernel.org>