tracing/eprobe: Add eprobe filter support
Add the filter option to the event probe. This is useful if
user wants to derive a new event based on the condition of
the original event.

E.g.

power:cpu_idle will occur before and after the actual idle state.
The event before one records with a state index, and the after one
has the state == 4294967295 ((u32)-1) on x86. Thus,


Signed-off-by: Masami Hiramatsu (Google) <mhiramat@kernel.org>