tracing: hist: Add poll support on hist file
Add poll syscall support on the `hist` file. The Waiter will be waken
up when the histogram is updated.

Currently, there is no way to wait for a specific event in userspace.
So user needs to peek the `trace` periodicaly, or wait on `trace_pipe`.
But that is not good idea to peek the `trace` for the event randomely
happens. And `trace_pipe` is not coming back until a page is filled
with events.
This allows user to wait for a specific events on `hist` file. User
can set a histogram trigger on the event which they want to monitor.
And epoll on its `hist` file.

NOTE: To read the hist file again, you must set the file offset to 0,
but just for monitoring the event, you may not need to read the
histogram.

Signed-off-by: Masami Hiramatsu (Google) <mhiramat@kernel.org>